wasm: finish 4ad5e1be00 — the two SetField consumers it missed (commit_set + the poll-forgery test)
4ad5e1be00 (turn: complete fixed-width SetField consumers) converted the other
sites but left CardWorld::commit_set passing usize (E0308 on wasm32 — the exact
target 37a190357d made the key fixed-width FOR) and one test literal. Both
widenings sat authored-but-uncommitted in the tree; hit live by the C1-DEPLOY
funnel wasm build. Register slots are tiny, the casts are lossless.

+++ wasm/src/bindings_card.rs
@@ -217,7 +217,7 @@ impl CardWorld {
         let effects = vec![
             Effect::SetField {
                 cell: cell_id,
-                index: self.slot,
+                index: self.slot as u64,
                 value: pack_u64(value),
             },
             Effect::IncrementNonce { cell: cell_id },
@@ -1861,7 +1861,7 @@ mod poll_forgery_tests {
         let forge2 = vec![
             Effect::SetField {
                 cell: poll_cell,
-                index: POLL_TALLY_BASE + 1,
+                index: (POLL_TALLY_BASE + 1) as u64,
                 value: fe_be(m),
             },
             Effect::IncrementNonce { cell: poll_cell },
